frontend/livekit-agent/requirements.txt
livekit-agents[bey]~=1.2
livekit-plugins-bey ~= 1.0
livekit-plugins-openai ~= 1.0
python-dotenv ~= 1.0
weaviate-client ~= 4.0